Give each claim-impound state its own preview under Wrapper::ClaimImpound
Also stops the markup digest from counting preview code: it only skipped
component_preview.rb and preview/, so a *_preview.rb sibling would have busted
the production fragment cache for every registration page on each edit.
