+++ spec/components/registrations/show/wrapper/component_system_spec.rb
@@ -7,19 +7,21 @@ RSpec.describe Registrations::Show::Wrapper::Component, :js, type: :system do
   # rest each need a preview, since reaching one otherwise means minting a token
   let(:carriers) { %w[prompt_chrome token_prompt wrapper] }
 
-  # Text that only appears once that alert rendered, keyed by the component's directory
-  # so the keys can be checked against what's on disk
-  let(:alert_text) do
-    # The card's heading is uppercased in CSS, so this anchors on its claim button
-    {"claim_impound" => "Claim found bike",
-     "claim_invitation" => "registered your bike on Bike Index",
-     "notification_token" => "Mark bike retrieved",
-     "recovery_prompt" => "Mark your bike recovered!",
-     "scanned_sticker" => "You scanned",
-     "sent_to_new_owner" => "You sent this"}
+  # The preview to visit for each alert, and text that only appears once it rendered.
+  # Keyed by the component's directory, so the keys can be checked against what's on
+  # disk — an alert with states of its own previews from a class per state, so this
+  # names the one that raises the card in its resting form
+  let(:alert_previews) do
+    # The claim card's heading is uppercased in CSS, so it anchors on its claim button
+    {"claim_impound" => ["claim_impound/with_stolen_registration/default", "Claim found bike"],
+     "claim_invitation" => ["component/claim_invitation", "registered your bike on Bike Index"],
+     "notification_token" => ["component/notification_token", "Mark bike retrieved"],
+     "recovery_prompt" => ["component/recovery_prompt", "Mark your bike recovered!"],
+     "scanned_sticker" => ["component/scanned_sticker", "You scanned"],
+     "sent_to_new_owner" => ["component/sent_to_new_owner", "You sent this"]}
   end
 
-  let(:preview_path) { "/rails/view_components/registrations/show/wrapper/component" }
+  let(:preview_path) { "/rails/view_components/registrations/show/wrapper" }
   let!(:organization) { FactoryBot.create(:organization_brakebills) }
   # ShowViews only offers the owner view to the owner or a superuser — without the seeded
   # superuser, sent_to_new_owner previews the public page
@@ -44,28 +46,34 @@ RSpec.describe Registrations::Show::Wrapper::Component, :js, type: :system do
       .select(&:directory?).map { |dir| dir.basename.to_s }.sort - carriers
   end
 
-  it "has a preview scenario for every kind of current alert" do
-    scenarios = Registrations::Show::Wrapper::ComponentPreview.public_instance_methods(false).map(&:to_s)
-
-    expect(alert_names).to eq alert_text.keys.sort
-    alert_names.each do |alert|
-      expect(scenarios).to include(alert), "CurrentAlerts::#{alert.camelize} has no preview scenario"
+  # Every registered preview, as the path the preview controller answers to
+  def preview_scenarios
+    ViewComponent::Preview.all.flat_map do |preview|
+      preview.public_instance_methods(false).map { |scenario| "#{preview.preview_name}/#{scenario}" }
     end
-    expect(scenarios).to include("no_overlay")
   end
 
-  it "renders the alert each scenario is named for, and none of them without one" do
-    alert_text.each do |scenario, text|
-      visit "#{preview_path}/#{scenario}"
+  it "has a preview for every kind of current alert" do
+    expect(alert_names).to eq alert_previews.keys.sort
+    alert_previews.each do |alert, (path, _)|
+      expect(preview_scenarios).to include("registrations/show/wrapper/#{path}"),
+        "CurrentAlerts::#{alert.camelize} has no preview"
+    end
+    expect(preview_scenarios).to include("registrations/show/wrapper/component/no_overlay")
+  end
+
+  it "renders the alert each preview is named for, and none of them without one" do
+    alert_previews.each_value do |(path, text)|
+      visit "#{preview_path}/#{path}"
 
       # The preview says so rather than raising when the record it needs is absent
       expect(page).to have_no_content("Nothing to preview")
       expect(page).to have_content(text)
     end
 
-    visit "#{preview_path}/no_overlay"
+    visit "#{preview_path}/component/no_overlay"
 
     expect(page).to have_no_content("Nothing to preview")
-    alert_text.each_value { |text| expect(page).to have_no_content(text) }
+    alert_previews.each_value { |(_, text)| expect(page).to have_no_content(text) }
   end
 end
